+++ ChangeLog.md
@@ -11,6 +11,7 @@ Next version
 ### Fixed
 - When an action assigned to a modal dialog's confirm button is disabled,
   properly indicate this on the button. (Also display the reason if available.)
+- Added `multiple` option to fileUpload
 
 ### Changed